+++ R/AllGenerics.R
@@ -98,7 +98,7 @@ setGeneric("addProcessHistory", function(object, ...)
 #'     parameter in \code{\link{profile-matrix}} documentation for more details.
 #'
 #' @param BPPARAM parallel processing setup. Defaults to `BPPARAM = bpparam()`.
-#'     See [bpparam()] for details.
+#'     See [BiocParallel::bpparam()] for details.
 #'
 #' @param centerSample \code{integer(1)} defining the index of the center sample
 #'     in the experiment. It defaults to
@@ -143,7 +143,7 @@ setGeneric("addProcessHistory", function(object, ...)
 #'
 #' @param family For `PeakGroupsParam`: `character(1)` defining the method for
 #'     loess smoothing. Allowed values are `"gaussian"` and `"symmetric"`. See
-#'     [loess()] for more information.
+#'     [stats::loess()] for more information.
 #'
 #' @param gapExtend For `ObiwarpParam`: `numeric(1)` defining the penalty for
 #'     gap enlargement. The default value for `gapExtend` depends on the value
@@ -177,8 +177,8 @@ setGeneric("addProcessHistory", function(object, ...)
 #' @param msLevel For `adjustRtime`: `integer(1)` defining the MS level on
 #'     which the alignment should be performed.
 #'
-#' @param object For `adjustRtime`: an [OnDiskMSnExp()], [XCMSnExp()],
-#'     [MsExperiment()] or [XcmsExperiment()] object.
+#' @param object For `adjustRtime`: an [MSnbase::OnDiskMSnExp()], [XCMSnExp()],
+#'     [MsExperiment::MsExperiment()] or [XcmsExperiment()] object.
 #'
 #' @param param The parameter object defining the alignment method (and its
 #'     setting).
@@ -212,7 +212,7 @@ setGeneric("addProcessHistory", function(object, ...)
 #'
 #' @param span For `PeakGroupsParam`: `numeric(1)` defining
 #'     the degree of smoothing (if `smooth = "loess"`). This parameter is
-#'     passed to the internal call to [loess()].
+#'     passed to the internal call to [stats::loess()].
 #'
 #' @param subset For `ObiwarpParam` and `PeakGroupsParam`: `integer` with the
 #'     indices of samples within the experiment on which the alignment models
@@ -463,7 +463,8 @@ setGeneric("chromPeakData<-", function(object, value)
 #'    The columns will be named as they are written in the `chromPeaks` object
 #'    with a prefix `"chrom_peak_"`. Defaults to `c("mz", "rt")`.
 #'
-#' @param BPPARAM parallel processing setup. Defaults to [bpparam()].
+#' @param BPPARAM parallel processing setup. Defaults to
+#'    [BiocParallel::bpparam()].
 #'
 #' @param ... ignored.
 #'
@@ -567,7 +568,8 @@ setGeneric("chromPeakSpectra", function(object, ...)
 #'   to a bell curve and the signal-to-noise ratio calculated on the residuals
 #'   of this test.
 #'
-#' @param BPPARAM Parallel processing setup. See [bpparam()] for details.
+#' @param BPPARAM Parallel processing setup. See
+#'     [BiocParallel::bpparam()] for details.
 #'
 #' @param chunkSize `integer(1)` defining the number of samples from which data
 #'     should be loaded and processed at a time.
@@ -701,8 +703,8 @@ setGeneric("family<-", function(object, value) standardGeneric("family<-"))
 #'     chromatogram.
 #'
 #' @param BPPARAM For `object` being an `XcmsExperiment`: parallel processing
-#'     setup. Defaults to `BPPARAM = bpparam()`. See [bpparam()] for more
-#'     information.
+#'     setup. Defaults to `BPPARAM = bpparam()`. See [BiocParallel::bpparam()]
+#'     for more information.
 #'
 #' @param chunkSize For `object` being an `XcmsExperiment`: `integer(1)`
 #'     defining the number of files from which the data should be loaded at
@@ -869,7 +871,8 @@ setGeneric("featureDefinitions<-", function(object, value)
 #' spectra per feature).
 #'
 #' The information from `featureDefinitions` for each feature can be included
-#' in the returned [Spectra()] object using the `featureColumns` parameter.
+#' in the returned [Spectra::Spectra()] object using the `featureColumns`
+#' parameter.
 #' This is useful for keeping details such as the median retention time (`rtmed`)
 #' or median m/z (`mzmed`). The columns will retain their names as specified
 #' in the `featureDefinitions` object, prefixed by `"feature_"`
@@ -878,9 +881,11 @@ setGeneric("featureDefinitions<-", function(object, value)
 #' as a metadata column named `"feature_id"`.
 #'
 #' See also [chromPeakSpectra()], as it supports a similar parameter for
-#' including columns from the chromatographic peaks in the returned spectra object.
+#' including columns from the chromatographic peaks in the returned spectra
+#' object.
 #' These parameters can be used in combination to include information from both
-#' the chromatographic peaks and the features in the returned [Spectra()].
+#' the chromatographic peaks and the features in the returned
+#' [Spectra::Spectra()].
 #' The *peak ID* (i.e., the row name of the peak in the `chromPeaks` matrix)
 #' is added as a metadata column named `"chrom_peak_id"`.
 #'
@@ -906,7 +911,8 @@ setGeneric("featureDefinitions<-", function(object, value)
 #'
 #' @return
 #'
-#' The function returns either a [Spectra()] (for `return.type = "Spectra"`)
+#' The function returns either a [Spectra::Spectra()] (for
+#' `return.type = "Spectra"`)
 #' or a `List` of `Spectra` (for `return.type = "List"`). For the latter,
 #' the order and the length matches parameter `features` (or if no `features`
 #' is defined the order of the features in `featureDefinitions(object)`).
@@ -1205,7 +1211,7 @@ setGeneric("filterFeatureDefinitions", function(object, ...)
 #'     object will remove previous results.
 #'
 #' @param BPPARAM Parallel processing setup. Uses by default the system-wide
-#'     default setup. See [bpparam()] for more details.
+#'     default setup. See [BiocParallel::bpparam()] for more details.
 #'
 #' @param chunkSize `integer(1)` for `object` being an `MsExperiment` or
 #'     [XcmsExperiment()]: defines the number of files (samples) for which the
@@ -1224,14 +1230,15 @@ setGeneric("filterFeatureDefinitions", function(object, ...)
 #'     will thus in most settings cause an out-of-memory error.
 #'     By setting `chunkSize = -1` the peak detection will be performed
 #'     separately, and in parallel, for each sample. This will however not work
-#'     for all `Spectra` *backends* (see eventually [Spectra()] for details).
+#'     for all `Spectra` *backends* (see eventually [Spectra::Spectra()] for
+#'     details).
 #'
 #' @param msLevel `integer(1)` defining the MS level on which the
 #'     chromatographic peak detection should be performed.
 #'
 #' @param object The data object on which to perform the peak detection. Can be
-#'     an [OnDiskMSnExp()], [XCMSnExp()], [MChromatograms()] or [MsExperiment()]
-#'     object.
+#'     an [MSnbase::OnDiskMSnExp()], [XCMSnExp()], [MSnbase::MChromatograms()]
+#'     or [MsExperiment::MsExperiment()] object.
 #'
 #' @param param The parameter object selecting and configuring the algorithm.
 #'
@@ -1301,7 +1308,8 @@ setGeneric("findChromPeaks", function(object, param, ...)
 #'     more information.
 #'
 #' @param BPPARAM if `object` is an `MsExperiment` or `XcmsExperiment`:
-#'     parallel processing setup. See [bpparam()] for more information.
+#'     parallel processing setup. See [BiocParallel::bpparam()] for more
+#'     information.
 #'
 #' @param ... currently not used.
 #'
@@ -1596,7 +1604,8 @@ setGeneric("loadRaw", function(object, ...) standardGeneric("loadRaw"))
 #' chromatographic peaks into features by providing their index in the
 #' object's `chromPeaks` matrix.
 #'
-#' @param BPPARAM parallel processing settings (see [bpparam()] for details).
+#' @param BPPARAM parallel processing settings (see [BiocParallel::bpparam()]
+#'     for details).
 #'
 #' @param chromPeaks For `manualChromPeaks`: `matrix` defining the boundaries
 #'     of the chromatographic peaks with one row per chromatographic peak and
@@ -1804,9 +1813,9 @@ setGeneric("rawMZ", function(object, ...) standardGeneric("rawMZ"))
 #'   Each MS2 chromatographic peak selected for an MS1 peak will thus represent
 #'   one **mass peak** in the reconstructed spectrum.
 #'
-#' The resulting [Spectra()] object provides also the peak IDs of the MS2
-#' chromatographic peaks for each spectrum as well as their correlation value
-#' with spectra variables *ms2_peak_id* and *ms2_peak_cor*.
+#' The resulting [Spectra::Spectra()] object provides also the peak IDs of
+#' the MS2 chromatographic peaks for each spectrum as well as their
+#' correlation value with spectra variables *ms2_peak_id* and *ms2_peak_cor*.
 #'
 #' @param object `XCMSnExp` with identified chromatographic peaks.
 #'
@@ -1833,8 +1842,8 @@ setGeneric("rawMZ", function(object, ...) standardGeneric("rawMZ"))
 #'     `chromPeaks`) of MS1 peaks for which MS2 spectra should be reconstructed.
 #'     By default they are reconstructed for all MS1 chromatographic peaks.
 #'
-#' @param BPPARAM parallel processing setup. See [bpparam()] for more
-#'     information.
+#' @param BPPARAM parallel processing setup. See [BiocParallel::bpparam()]
+#'     for more information.
 #'
 #' @param return.type `character(1)` defining the type of the returned object.
 #'     Only `return.type = "Spectra"` is supported, `return.type = "MSpectra"`
@@ -1844,14 +1853,14 @@ setGeneric("rawMZ", function(object, ...) standardGeneric("rawMZ"))
 #'
 #' @return
 #'
-#' - [Spectra()] object (defined in the `Spectra` package) with the
+#' - [Spectra::Spectra()] object (defined in the `Spectra` package) with the
 #'   reconstructed MS2 spectra for all MS1 peaks in `object`. Contains
 #'   empty spectra (i.e. without m/z and intensity values) for MS1 peaks for
 #'   which reconstruction was not possible (either no MS2 signal was recorded
 #'   or the correlation of the MS2 chromatographic peaks with the MS1
 #'   chromatographic peak was below threshold `minCor`. Spectra variables
-#'   `"ms2_peak_id"` and `"ms2_peak_cor"` (of type [CharacterList()]
-#'   and [NumericList()] with length equal to the number of peaks per
+#'   `"ms2_peak_id"` and `"ms2_peak_cor"` (of type [IRanges::CharacterList()]
+#'   and [IRanges::NumericList()] with length equal to the number of peaks per
 #'   reconstructed MS2 spectrum) providing the IDs and the correlation of the
 #'   MS2 chromatographic peaks from which the MS2 spectrum was reconstructed.
 #'   As retention time the median retention times of all MS2 chromatographic
@@ -1947,7 +1956,7 @@ setGeneric("reconstructChromPeakSpectra", function(object, ...)
 #'
 #' @param BPPARAM parameter object to set up parallel processing. Uses the
 #'     default parallel processing setup returned by `bpparam()`. See
-#'     [bpparam()] for details and examples.
+#'     [BiocParallel::bpparam()] for details and examples.
 #'
 #' @param chunkSize For `refineChromPeaks` if `object` is either an
 #'     `XcmsExperiment`: `integer(1)` defining the number of files (samples)